requirements.txt
langchain>=0.0.275
langchain-community>=0.0.15
streamlit>=1.25.0
python-dotenv>=1.0.0
requests>=2.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
notebook>=7.0.0
ipykernel>=6.0.0
pypdf